requirements.txt
alembic
python-dotenv
psycopg2-binary
loguru
asyncpg
pandas
sqlalchemy